Made the app unusable on small viewport devices
Hopefully I learn Responsive Design some day, for now, the app will no longer work on small viewport devices (anything with width < 1800px). this is made to stop the app being used especially on phones, which, currently, is a horrible idea.
This effect is app-wide, as it's embedded in AuthWrapper, rather that "/" or "/rerouter".