Dropped PlanetScale for JawsDB
Back to JawsDB and Amazon RDS for database hosting.
Turns out PlanetScale is the worst for non-production projects.
If your DB doesn't get a WRITE operation for a while, PlanetScale puts it to "sleep", which is another way of saying that they're shutting down the database (You'd no longer be able to connect to your database), and it must be restarted in the web dashboard for it to be connect-able again. To avoid this, you HAVE to perform some sort of WRITE operation on your app every few weeks. This sucks.
Never using PlanetScale again.