Remove User object from localStorage on signOut
Signout now removes the User object stored in localStorage with the key "recoil-persist". Now on signout, the recoil-persist key is removed from localStorage.
"/api/signoutAPI" is now "/api/signout".
Changed root path ( "/" ) behavior to just reroute to either the login page or straight into the app ("/buildings") according to if the user object is not empty (has attributes).